# specs/003-conciliacao-fiscal/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Auditoria-JB - API de Conciliação Fiscal
  version: 1.0.0
  description: >-
    Contrato da sexta ferramenta. A autenticação usa o cookie de sessão já
    existente no Auditoria-JB. Respostas 403 também entram na trilha geral.
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/jobs/{job_id}:
    get:
      tags: [Conciliação]
      operationId: obterJob
      summary: Acompanha um job de processamento (rota compartilhada)
      description: >-
        Rota genérica já existente no Auditoria-JB, compartilhada pelas seis
        ferramentas. Documentada aqui porque /processar devolve `job_id` e o
        contrato ficaria incompleto sem o endpoint que o consome.

        Exige apenas login: o controle de acesso é a PROPRIEDADE do job, não
        uma permissão de ação. Job de outro usuário responde 404 com o mesmo
        corpo de um job inexistente — um 403 confirmaria que aquele ID existe.

        `ferramenta` é obrigatória e amarra o job à aba que o criou (FR-005 e
        SEC003): um job_id de outra aba responde como inexistente, mesmo sendo
        do próprio usuário. Omiti-la é erro de contrato (422), nunca um atalho
        silencioso. Ferramenta desconhecida cai no mesmo 404 — a rota jamais
        revela se o problema foi o ID, o dono ou a aba.

        Fora da trilha geral de propósito: o front consulta a cada 700 ms e
        auditar isso afogaria o que importa.
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/FerramentaQuery'
      responses:
        '200':
          description: >-
            Estado do job. Concluído o processamento da conciliação,
            `resultado` é um ImportBatchResult; o mesmo objeto continua
            disponível de forma persistente em /api/conciliacao/lotes/{lote_id}.
          content:
            application/json:
              schema:
                type: object
                required: [status, erro, resultado, descricao]
                properties:
                  status:
                    type: string
                    enum: [executando, concluido, erro]
                  erro: { type: string }
                  descricao: { type: string }
                  resultado:
                    oneOf:
                      - { $ref: '#/components/schemas/ImportBatchResult' }
                      - { type: 'null' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auditoria_sessao
  parameters:
...
    JobId:
      name: job_id
      in: path
      required: true
      schema: { type: string }
    FerramentaQuery:
      name: ferramenta
      in: query
      required: true
      description: >-
        Aba que criou o recurso. Valor divergente do job, ou desconhecido,
        responde 404 — igual a um ID inexistente.
      schema:
        type: string
        enum: [comparador, diff, conferencia, extracao, produtos, conciliacao]
...
  responses:
    Unauthorized:
      description: Login necessário
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Recurso inexistente ou não pertencente ao usuário
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Entrada inválida ou regra de negócio não atendida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code: { type: string }
        current_revision: { type: integer }
...
    ImportItemResult:
      type: object
      required: [id, nome_recebido, resultado, mensagem]
      properties:
        id: { type: integer }
        fonte_id: { type: [integer, 'null'] }
        conciliacao_id: { type: [integer, 'null'] }
        nome_recebido: { type: string }
        sha256: { type: [string, 'null'], pattern: '^[a-f0-9]{64}$' }
        resultado:
          type: string
          enum: [processado, duplicado, conflitante, rejeitado]
        mensagem: { type: string }
        codigo: { type: [string, 'null'] }
    ImportBatchResult:
      type: object
      required:
        - id
        - estado
        - total_arquivos
        - processados
        - duplicados
        - conflitantes
        - rejeitados
        - itens
      properties:
        id: { type: integer }
        estado:
          type: string
          enum: [recebido, processando, concluido, interrompido]
        total_arquivos: { type: integer, minimum: 0 }
        processados: { type: integer, minimum: 0 }
        duplicados: { type: integer, minimum: 0 }
        conflitantes: { type: integer, minimum: 0 }
        rejeitados: { type: integer, minimum: 0 }
        criado_em: { type: string, format: date-time }
        iniciado_em: { type: [string, 'null'], format: date-time }
        concluido_em: { type: [string, 'null'], format: date-time }
        itens:
          type: array
          items: { $ref: '#/components/schemas/ImportItemResult' }
...